Refactored updates command and handlers
