+++ Library.LibrarianBot/Commands/Updates/OnCallbackQueryUpdateCommand.cs
@@ -1,0 +1,7 @@
+﻿namespace Library.LibrarianBot.Commands.Updates
+{
+    public class OnCallbackQueryUpdateCommand : UpdatesCommandBase
+    {
+        
+    }
+}
