+++ Library.LibrarianBot/Commands/Updates/OnEditedMessageUpdateCommand.cs
@@ -1,0 +1,7 @@
+﻿namespace Library.LibrarianBot.Commands.Updates
+{
+    public class OnEditedMessageUpdateCommand : UpdatesCommandBase
+    {
+        
+    }
+}
