+++ Library.LibrarianBot/Commands/Updates/UpdatesCommandBase.cs
@@ -3,7 +3,7 @@ using Telegram.Bot.Types;
 
 namespace Library.LibrarianBot.Commands.Updates
 {
-    public class OnCallbackQueryReceivedCommand : IRequest<bool>
+    public class UpdatesCommandBase : IRequest<bool>
     {
         public Message Message { get; init; }
     }
