+++ Library.LibrarianBot/Handlers/Updates/OnMessageUpdateCommandHandler.cs
@@ -6,20 +6,20 @@ using Telegram.Bot;
 
 namespace Library.LibrarianBot.Handlers.Updates
 {
-    public class OnMessageReceivedCommandHandler : IRequestHandler<OnMessageReceivedCommand, bool>
+    public class OnMessageUpdateCommandHandler : IRequestHandler<OnMessageUpdateCommand, bool>
     {
         private readonly ITelegramBotClient _botClient;
         private readonly IAppDBContext _ctx;
         private readonly ISender _sender;
 
-        public OnMessageReceivedCommandHandler(ITelegramBotClient botClient, IAppDBContext ctx ,ISender sender)
+        public OnMessageUpdateCommandHandler(ITelegramBotClient botClient, IAppDBContext ctx ,ISender sender)
         {
             _botClient = botClient;
             _ctx = ctx;
             _sender = sender;
         }
 
-        public async Task<bool> Handle(OnMessageReceivedCommand request, CancellationToken cancellationToken)
+        public async Task<bool> Handle(OnMessageUpdateCommand request, CancellationToken cancellationToken)
         {
             bool result = request.Message.Text! switch
             {
